Revert "adds godoc"
This reverts commit d58de406aeecd52ccac54538e6d47508e9e504ae.

+++ gossipsub.go
@@ -223,14 +223,14 @@ func NewGossipSubWithRouter(ctx context.Context, h host.Host, rt PubSubRouter, o
 type ReadOnlyRoutingMetadata struct {
 	mu sync.RWMutex
 	// metda-date will be updated every k-heartbeats by the router. This is used to avoid unnecessary updates.
-	k uint64
+	k int
 	// heartbeat counter
-	count  uint64
+	count  int
 	mesh   map[string]map[peer.ID]struct{}
 	fanout map[string]map[peer.ID]struct{}
 }
 
-func newReadOnlyRoutingMetadataTracer(k uint64) *ReadOnlyRoutingMetadata {
+func newReadOnlyRoutingMetadataTracer(k int) *ReadOnlyRoutingMetadata {
 	return &ReadOnlyRoutingMetadata{
 		k:      k,
 		count:  0,
@@ -288,9 +288,7 @@ func (r *ReadOnlyRoutingMetadata) update(mesh map[string]map[peer.ID]struct{}, f
 
 type GossipSubRouterOption func(*GossipSubRouter)
 
-// WithReadOnlyRoutingMetadata returns a GossipSubRouterOption that enables the read-only routing metadataTracer.
-// The metadataTracer is updated every k-heartbeats by the router.
-func WithReadOnlyRoutingMetadata(k uint64) GossipSubRouterOption {
+func WithReadOnlyRoutingMetadata(k int) GossipSubRouterOption {
 	return func(r *GossipSubRouter) {
 		r.metadataTracer = newReadOnlyRoutingMetadataTracer(k)
 	}
